+++ src/seedcase_sprout/__init__.py
@@ -39,6 +39,7 @@ from .properties import (
 )
 from .read_properties import read_properties
 from .read_resource_batches import read_resource_batches
+from .read_resource_data import read_resource_data
 from .write_file import write_file
 from .write_properties import write_properties
 from .write_resource_batch import write_resource_batch
@@ -72,6 +73,7 @@ __all__ = [
     "extract_resource_properties",
     "join_resource_batches",
     "read_resource_batches",
+    "read_resource_data",
     "write_resource_batch",
     "create_resource_properties_script",
     "write_resource_data",